+++ App.js
@@ -1,21 +1,27 @@
 import React, { useState } from "react";
-import { StyleSheet, SafeAreaView, StatusBar, Image } from "react-native";
-import { createBottomTabNavigator } from "@react-navigation/bottom-tabs";
+import { StyleSheet, SafeAreaView, StatusBar, Text } from "react-native";
+import { createStackNavigator } from "@react-navigation/stack";
 import { NavigationContainer, DefaultTheme } from "@react-navigation/native";
 import AppLoading from "expo-app-loading";
 import * as Font from "expo-font";
 
 // Assets
-import { Colors, Metrics } from "./assets/Themes";
-import CustomIcons from "./assets/Fonts";
+import { Colors } from "./assets/Themes";
 
 // Components
-import { Browse, Upload, Profile, Saved, Notifications } from "./screens";
+import { Main } from "./screens";
+
+// const Main = () => {
+//   return (
+//     <SafeAreaView>
+//       <Text>Hi</Text>
+//     </SafeAreaView>
+//   );
+// };
 
 // Use this to get colors from theme
 // import { useTheme } from '@react-navigation/native';
 // const { colors } = useTheme();
-
 const Theme = {
   ...DefaultTheme,
   colors: {
@@ -27,7 +33,7 @@ const Theme = {
   },
 };
 
-const Tab = createBottomTabNavigator();
+const Stack = createStackNavigator();
 
 export default function App() {
   const [appLoading, setAppLoading] = useState(false);
@@ -66,87 +72,13 @@ export default function App() {
       <SafeAreaView style={styles.bottomSafeArea}>
         <StatusBar barStyle="light-content" />
         <NavigationContainer theme={Theme}>
-          <Tab.Navigator
-            screenOptions={({ route }) => ({
-              tabBarIcon: ({ focused, color, size }) => {
-                let icon;
-                switch (route.name) {
-                  case "Browse":
-                    icon = focused ? (
-                      <CustomIcons
-                        name="home-filled"
-                        color={color}
-                        size={size}
-                      />
-                    ) : (
-                      <CustomIcons name="home" color={color} size={size} />
-                    );
-                    break;
-                  case "Saved":
-                    icon = focused ? (
-                      <CustomIcons name="bookmark" color={color} size={size} />
-                    ) : (
-                      <CustomIcons name="bookmark" color={color} size={size} />
-                    );
-                    break;
-                  case "Upload":
-                    icon = focused ? (
-                      <Image
-                        source={require("./assets/Images/add.png")}
-                        style={styles.upload}
-                      />
-                    ) : (
-                      <Image
-                        source={require("./assets/Images/add.png")}
-                        style={styles.upload}
-                      />
-                    );
-                    break;
-                  case "Notifications":
-                    icon = focused ? (
-                      <CustomIcons
-                        name="notification"
-                        color={color}
-                        size={size}
-                      />
-                    ) : (
-                      <CustomIcons
-                        name="notification"
-                        color={color}
-                        size={size}
-                      />
-                    );
-                    break;
-                  case "Profile":
-                    icon = focused ? (
-                      <CustomIcons name="user" color={color} size={size} />
-                    ) : (
-                      <CustomIcons name="user" color={color} size={size} />
-                    );
-                    break;
-                  default:
-                    null;
-                }
-                return icon;
-              },
-            })}
-            tabBarOptions={{
-              activeTintColor: Colors.white,
-              inactiveTintColor: "gray",
-              showLabel: false,
-              style: {
-                backgroundColor: Colors.blur,
-                height: Metrics.tabBarHeight,
-                borderTopWidth: 0,
-              },
-            }}
-          >
-            <Tab.Screen name="Browse" component={Browse} />
-            <Tab.Screen name="Saved" component={Saved} />
-            <Tab.Screen name="Upload" component={Upload} />
-            <Tab.Screen name="Notifications" component={Notifications} />
-            <Tab.Screen name="Profile" component={Profile} />
-          </Tab.Navigator>
+          <Stack.Navigator initialRouteName="Main">
+            <Stack.Screen
+              name="Main"
+              component={Main}
+              options={{ headerShown: false }}
+            />
+          </Stack.Navigator>
         </NavigationContainer>
       </SafeAreaView>
     </>
@@ -162,11 +94,5 @@ const styles = StyleSheet.create({
     flex: 1,
     borderRadius: 21,
     backgroundColor: Colors.background1,
-  },
-  upload: {
-    marginTop: 10,
-    width: 60,
-    height: undefined,
-    aspectRatio: 1,
   },
 });
