+++ assets/Themes/Images.js
@@ -1,6 +1,7 @@
 const images = {
   logo: require("../Images/logo.png"),
   homeIcon: require("../Images/home-icon.svg"),
+  tameImpala: require("../Images/tame_impala.png"),
 };
 
 export default images;
