+++ components/Bubble/Bubble.js
@@ -1,11 +1,15 @@
 import React from "react";
-import { View, Text } from "react-native";
+import { View, Text, Image } from "react-native";
+
+// Assets
+import { Images } from "../../assets/Themes";
 
 export function Bubble() {
   return (
     <View>
       {/* Three Circles - image, black, genre color */}
       <Text>Bubble</Text>
+      <Image source={Images.tameImpala} />
     </View>
   );
 }
