+++ package.json
@@ -8,6 +8,7 @@
     "eject": "expo eject"
   },
   "dependencies": {
+    "@react-native-community/masked-view": "^0.1.10",
     "@react-navigation/bottom-tabs": "^5.11.7",
     "@react-navigation/native": "^5.9.2",
     "@react-navigation/stack": "^5.14.2",
@@ -18,6 +19,7 @@
     "react-dom": "16.13.1",
     "react-iconly": "^1.1.3",
     "react-native": "https://github.com/expo/react-native/archive/sdk-40.0.1.tar.gz",
+    "react-native-gesture-handler": "^1.10.1",
     "react-native-screens": "^2.17.1",
     "react-native-vector-icons": "^8.0.0",
     "react-native-web": "~0.13.12"
