+++ screens/index.js
@@ -3,3 +3,4 @@ export { default as Upload } from "./Upload";
 export { default as Profile } from "./Profile";
 export { default as Saved } from "./Saved";
 export { default as Notifications } from "./Notifications";
+export { default as Main } from "./Main";
